World Building & Block Collision : Ok

+++ include/Monster.hpp
@@ -22,8 +22,8 @@ enum {
 
 const int MONSTER_WIDTH = 200;
 const int MONSTER_HEIGHT = 200;
-const float MONSTER_WALK_SPEED = 0.2f;
-const float MONSTER_RUN_SPEED = 0.6f;
+const double MONSTER_WALK_SPEED = 0.2f;
+const double MONSTER_RUN_SPEED = 0.6f;
 
 class Monster : public AnimatedEntity {
